+++ backend/src/main/java/mk/com/store/games/gamestore/service/inplementation/PublisherServiceImplementation.java
@@ -40,40 +40,33 @@ public class PublisherServiceImplementation  implements PublisherService{
 
     @Override
     public List<Publisher> getAllPublisherByUser(String username) throws UserNotFoundException {
+        if(username == null) throw new IllegalArgumentException("Username can not be null");
+        if(username.trim().equals("")) throw new IllegalArgumentException("Username can not be empty string");
+        if(username.contains(" ")) throw new IllegalArgumentException("Username can not contain empty spaces");
         User user = userRepository.findByUsername(username).stream().findFirst().orElseThrow(UserNotFoundException::new);
         return user.getPublishers();
     }
 
     @Override
     public Optional<Publisher> addPublisher(PublisherDto publisherDto) throws UserNotFoundException {
+        if(publisherDto == null) throw new IllegalArgumentException("Publisher info can not be null");
+        if(publisherDto.getUsername() == null) throw new IllegalArgumentException("Username can not be null");
+        if(publisherDto.getUsername().trim().equals("")) throw new IllegalArgumentException("Username can not be empty string");
+        if(publisherDto.getUsername().contains(" ")) throw new IllegalArgumentException("Username can not contain empty spaces");
         User user = userRepository.findByUsername(publisherDto.getUsername()).stream().findFirst().orElseThrow(UserNotFoundException::new);
         Publisher publisher = new Publisher(publisherDto.getName(), publisherDto.getDescription());
-        publisher = publisherRepository.save(publisher);
+        publisherRepository.save(publisher);
         user.getPublishers().add(publisher);
         userRepository.save(user);
         return Optional.of(publisher);
     }
 
     @Override
-    public Optional<Publisher> addEditor(String publisherId, String username) throws UserNotFoundException, PublisherNotFoundException {
-        User user = userRepository.findByUsername(username).stream().findFirst().orElseThrow(UserNotFoundException::new);
-        Publisher publisher = publisherRepository.findById(publisherId).orElseThrow(PublisherNotFoundException::new);
-        user.getPublishers().add(publisher);
-        userRepository.save(user);
-        return Optional.of(publisherRepository.save(publisher));
-    }
-
-    @Override
-    public Optional<Publisher> removeEditor(String publisherId, String username) throws UserNotFoundException, PublisherNotFoundException {
-        User user = userRepository.findByUsername(username).stream().findFirst().orElseThrow(UserNotFoundException::new);
-        Publisher publisher = publisherRepository.findById(publisherId).orElseThrow(PublisherNotFoundException::new);
-        user.getPublishers().remove(publisher);
-        userRepository.save(user);
-        return Optional.of(publisherRepository.save(publisher));
-    }
-
-    @Override
-    public Optional<Publisher> editPublisher(String publisherId, PublisherDto publisherDto) throws UserNotFoundException, PublisherNotFoundException {
+    public Optional<Publisher> editPublisher(String publisherId, PublisherDto publisherDto) throws PublisherNotFoundException {
+        if(publisherDto == null) throw new IllegalArgumentException("Publisher info can not be null");
+        if(publisherId == null) throw new IllegalArgumentException("PublisherId can not be null");
+        if(publisherId.trim().equals("")) throw new IllegalArgumentException("publisherId can not be empty string");
+        if(publisherId.contains(" ")) throw new IllegalArgumentException("publisherId can not contain empty spaces");
         Publisher publisher = publisherRepository.findById(publisherId).orElseThrow(PublisherNotFoundException::new);
         publisher.setName(publisherDto.getName());
         publisher.setDescription(publisherDto.getDescription());
@@ -81,8 +74,11 @@ public class PublisherServiceImplementation  implements PublisherService{
     }
 
     @Override
-    public Optional<Boolean> removePublisher(String publisherId) throws UserNotFoundException, PublisherNotFoundException {
-        Publisher publisher = publisherRepository.findById(publisherId).stream().filter(pub -> pub.getId().equals(publisherId)).findFirst().orElseThrow(PublisherNotFoundException::new);
+    public Optional<Boolean> removePublisher(String publisherId) throws PublisherNotFoundException {
+        if(publisherId == null) throw new IllegalArgumentException("PublisherId can not be null");
+        if(publisherId.trim().equals("")) throw new IllegalArgumentException("PublisherId can not be empty string");
+        if(publisherId.contains(" ")) throw new IllegalArgumentException("PublisherId can not contain empty spaces");
+        Publisher publisher = publisherRepository.findById(publisherId).orElseThrow(PublisherNotFoundException::new);
         publisher.getStudios().forEach(studio -> {
             try {
                 developerService.removeDeveloper(studio.getId());
